Refactor domain, move cursor and features to own files
