+++ ogc/features/domain/features.go
@@ -9,6 +9,45 @@ import (
 	"github.com/jmoiron/sqlx"
 )
 
+// featureCollectionType allows the GeoJSON type to be automatically set during json marshalling
+type featureCollectionType struct{}
+
+func (fc *featureCollectionType) MarshalJSON() ([]byte, error) {
+	return []byte(`"FeatureCollection"`), nil
+}
+
+func (fc *featureCollectionType) UnmarshalJSON([]byte) error { return nil }
+
+// FeatureCollection is a GeoJSON FeatureCollection with extras such as links
+type FeatureCollection struct {
+	Links []Link `json:"links,omitempty"`
+
+	NumberReturned int                   `json:"numberReturned"`
+	Type           featureCollectionType `json:"type"`
+	Features       []*Feature            `json:"features"`
+}
+
+// Feature is a GeoJSON Feature with extras such as links
+type Feature struct {
+	// we overwrite ID since we want to make it a required attribute. We also expect feature ids to be
+	// auto-incrementing integers (which is the default in geopackages) since we use it for cursor-based pagination.
+	ID    int64  `json:"id"`
+	Links []Link `json:"links,omitempty"`
+
+	geojson.Feature
+}
+
+// Link according to RFC 8288, https://datatracker.ietf.org/doc/html/rfc8288
+type Link struct {
+	Length    int64  `json:"length,omitempty"`
+	Rel       string `json:"rel"`
+	Title     string `json:"title,omitempty"`
+	Type      string `json:"type,omitempty"`
+	Href      string `json:"href"`
+	Hreflang  string `json:"hreflang,omitempty"`
+	Templated bool   `json:"templated,omitempty"`
+}
+
 // MapRowsToFeatures datasource agnostic mapper from SQL rows/result set to Features domain model
 func MapRowsToFeatures(rows *sqlx.Rows, fidColumn string, geomColumn string,
 	geomMapper func([]byte) (geom.Geometry, error)) ([]*Feature, error) {
